* Konfigurert appen til å koble seg til Kafka i preprod. * Appen lytter på en informasjon-topic. * Har et midlertidig endepunkt som gjør at man kan produsere et nytt informasjon-event, som man kan se at appen leser inn igjen fra Kafka.
