+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/App.kt
@@ -1,6 +1,9 @@
 package no.nav.personbruker.dittnav.eventaggregator
 
 import kotlinx.coroutines.runBlocking
+import no.nav.personbruker.dittnav.eventaggregator.config.Config
+import no.nav.personbruker.dittnav.eventaggregator.config.Config.informasjonTopicName
+import no.nav.personbruker.dittnav.eventaggregator.config.Environment
 import no.nav.personbruker.dittnav.skjema.Informasjon
 
 fun main(args: Array<String>) {
@@ -9,7 +12,7 @@ fun main(args: Array<String>) {
         Server.startServer(System.getenv("PORT")?.toInt() ?: 8080).start()
         lesInnProperties()
 
-        startConsumer()
+        startInformasjonConsumer()
     }
 
 }
@@ -31,9 +34,9 @@ private fun lesInnIkkeSensitivPropertyDefinertINaisManifiestet() {
     println("Leste inn miljøvariabelen IKKE_SENSITIV_PROPERTY til å være: $propertyValue")
 }
 
-private fun startConsumer() {
+private fun startInformasjonConsumer() {
     Consumer.apply {
-        create(topics = listOf("example.topic.dittnav.informasjon"), kafkaProps = Config.consumerProps(Environment()))
+        create(topics = listOf(informasjonTopicName), kafkaProps = Config.consumerProps(Environment()))
         fetchFromKafka<Informasjon>()
     }
 }
