+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/api/Produce.kt
@@ -16,13 +16,13 @@ fun Routing.produceEventsApi() {
     }
 
     post("/produce/oppgave") {
-        Server.producer.produceOppgaveEvent()
+//        Server.producer.produceOppgaveEvent() // Midlertidig kommetert ut, fram til topic-en er tilgjengelig
         val msg = "Produced Oppgave-event"
         call.respond(HttpStatusCode.OK, msg)
     }
 
     post("/produce/melding") {
-        Server.producer.produceMeldingEvent()
+//        Server.producer.produceMeldingEvent() // Midlertidig kommetert ut, fram til topic-en er tilgjengelig
         val msg = "Produced Melding-event"
         call.respond(HttpStatusCode.OK, msg)
     }
