+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/config/Environment.kt
@@ -1,0 +1,13 @@
+package no.nav.personbruker.dittnav.eventaggregator.config
+
+data class Environment(val bootstrapServers: String = getEnvVar("KAFKA_BOOTSTRAP_SERVERS", "localhost:9092"),
+                       val schemaRegistryUrl: String = getEnvVar("KAFKA_SCHEMAREGISTRY_SERVERS", "http://localhost:8081"),
+                       val username: String = getEnvVar("FSS_SYSTEMUSER_USERNAME", "username"),
+                       val password: String = getEnvVar("FSS_SYSTEMUSER_PASSWORD", "password"),
+                       val groupId: String = getEnvVar("GROUP_ID", "dittnav_events")
+)
+
+fun getEnvVar(varName: String, defaultValue: String? = null): String {
+    return System.getenv(varName) ?: defaultValue
+    ?: throw IllegalArgumentException("Variable $varName cannot be empty")
+}
